Updated Stylecop to the latest prerelease build
There's a known bug in the latest stable version of stylecop which raises a warning when using "record" types, which is fixed in the prerelease version
The warning seems to cause our pipeline to fail, so I've switched to the prerelease version for now.
We should switch back when a new stable version is released.